% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gastro_label_binary.R
\docType{data}
\name{gastro_label_binary}
\alias{gastro_label_binary}
\title{Gastrointestinal binary labels}
\description{
A panel of seven endoscopists viewed the videos and determined which patient needs resection (malignant) or no-resection  (benign).
}
\references{
\url{http://www.depeca.uah.es/colonoscopy_dataset/}
}
