% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gastro_label_trinary.R
\docType{data}
\name{gastro_label_trinary}
\alias{gastro_label_trinary}
\title{Gastrointestinal trinary labels}
\description{
Gastrointestinal trinary ground truth (Adenoma,  Serrated, and Hyperplastic)
}
\references{
\url{http://www.depeca.uah.es/colonoscopy_dataset/}
}
