% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par2list.R
\name{par2list}
\alias{par2list}
\title{Transfer a vector into a list}
\usage{
par2list(par, g, p, ncov = 2, type = c("ign", "full"))
}
\arguments{
\item{par}{A vector with list information.}

\item{g}{Number of multivariate Gaussian groups.}

\item{p}{Dimension of observation vecor.}

\item{ncov}{Options of structure of sigma matrix;  the default value is 2;
\code{ncov} = 1 for a common covariance matrix that \code{sigma} is a \eqn{p\times p} matrix.
\code{ncov} = 2 for the unequal  covariance/scale matrices that
\code{sigma} represents a list of g matrices with dimension \eqn{p\times p \times g}.}

\item{type}{Two types to fit to the model, 'ign' indicates fitting the model on the basis of the missing-label mechanism ignored,and 'full' indicates fitting the model on the basis of the missing-label mechanism.}
}
\value{
\item{parlist}{Return a list including \code{mu}, \code{pi}, \code{sigma} and \code{xi}.}
}
\description{
Transfer a vector into a list
}
