% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2par.R
\name{list2par}
\alias{list2par}
\title{Transfer a list into a vector}
\usage{
list2par(
  p,
  g,
  pi,
  mu,
  sigma,
  ncov = 2,
  xi = NULL,
  type = c("ign", "full", "com")
)
}
\arguments{
\item{p}{Dimension of observation vecor.}

\item{g}{Number of multivariate normal classes.}

\item{pi}{A g-dimensional vector for the initial values of the mixing proportions.}

\item{mu}{A \eqn{p \times g} matrix for the initial values of the location parameters.}

\item{sigma}{A \eqn{p\times p} covariance matrix if \code{ncov=1}, or a list of g covariance matrices with dimension \eqn{p\times p \times g} if \code{ncov=2}.}

\item{ncov}{Options of structure of sigma matrix;  the default value is 2;
\code{ncov} = 1 for a common covariance matrix;
\code{ncov} = 2 for the unequal  covariance/scale matrices.}

\item{xi}{A 2-dimensional vector containing the initial values of the coefficients in the logistic function of the Shannon entropy.}

\item{type}{Three types to fit to the model, 'ign' indicates fitting the model on the basis of the likelihood that ignores the missing label mechanism,
'full' indicates that the model to be fitted on the basis of the full likelihood, taking into account the missing-label mechanism,
and 'com' indicate that the model to be fitted to a completed classified sample.}
}
\value{
\item{par}{a vector including all list information}
}
\description{
Transfer a list into a vector
}
